refactor(core): split skills/MCP/mirror configs into types::skills_mcp submodule
Continues the `core/src/types.rs` kitchen-sink split. Moves `SkillsConfig`,
`McpConfig`, `McpServerEntry`, `McpTransportConfig`, `McpTrustLevel`, and
the legacy `MirrorConfig` out of the flat root namespace into
`assistant_core::types::skills_mcp::*`, dropping their entries from the
`lib.rs` re-export.

`types.rs`:
- `pub mod skills_mcp;` declared at the migration site.
- `use self::skills_mcp::{McpConfig, MirrorConfig, SkillsConfig};` added
  so `AssistantConfig` struct fields keep resolving.
- Test module gains `use super::skills_mcp::{McpTransportConfig, McpTrustLevel};`
  since the existing serde-roundtrip tests reference them.
- Dropped now-unused `std::collections::HashMap` import (it was only used
  by `McpServerEntry::env`, which moved out).

Call-site changes (2 files):
- `crates/mcp-client/src/manager.rs` — `McpServerEntry`, `McpTransportConfig`,
  `McpTrustLevel` from `types::skills_mcp`.
- `crates/interface-cli/src/cmd_doctor.rs` — `McpTransportConfig` split out
  of a combined `assistant_core::{...}` import.

Tests:
- `cargo test -p assistant-core --lib` — 183 pass
- `cargo test -p assistant-mcp-client --lib` — 11 pass
- `cargo clippy --workspace -- -D warnings` — green
