+++ crates/interface-cli/src/cmd_doctor.rs
@@ -8,8 +8,9 @@ use std::path::{Path, PathBuf};
 
 use anyhow::Result;
 
+use assistant_core::types::skills_mcp::McpTransportConfig;
 use assistant_core::types::storage::BusKind;
-use assistant_core::{AssistantConfig, LlmProviderKind, McpTransportConfig};
+use assistant_core::{AssistantConfig, LlmProviderKind};
 
 // ── Helpers ──────────────────────────────────────────────────────────────────
